LayoutTests/imported/w3c/web-platform-tests/svg/types/scripted/SVGAnimatedEnumeration-SVGMarkerElement-expected.txt


FAIL Use of SVGAnimatedEnumeration within SVGMarkerElement assert_equals: expected 0 but got 10
